pyproject.toml
[project]
name = "trinity-counselor"
version = "0.1.0"
description = "Trinity: Multi-Agent AI Relationship Counseling System"
requires-python = ">=3.11"

dependencies = [
    # Web framework
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.9",

    # LangChain + LangGraph
    "langchain>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-anthropic>=0.1.15",
    "langchain-openai>=0.1.8",       # embeddings (text-embedding-3-small)
    "langchain-postgres>=0.0.9",     # PGVector store
    "langgraph>=0.2.0",
    "langgraph-checkpoint-postgres>=1.0.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "psycopg[binary,pool]>=3.1.19",  # psycopg3 for langgraph checkpointer
    "alembic>=1.13.0",

    # Neo4j
    "neo4j>=5.20.0",

    # Auth & Security
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "cryptography>=42.0.0",

    # Config
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",

    # Utilities
    "httpx>=0.27.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
